# *: reduce useless `promtheus.WithLabelValues` call