--- a/sessionctx/stmtctx/stmtctx.go
+++ b/sessionctx/stmtctx/stmtctx.go
@@ -15,6 +15,7 @@ package stmtctx
 
 import (
 	"math"
+	"strconv"
 	"sync"
 	"time"
 
@@ -93,6 +94,7 @@ type StatementContext struct {
 
 		message           string
 		warnings          []SQLWarn
+		errorCount        uint16
 		histogramsNotLoad bool
 		execDetails       execdetails.ExecDetails
 	}
@@ -265,31 +267,43 @@ func (sc *StatementContext) WarningCount() uint16 {
 	return wc
 }
 
-// NumWarnings gets warning count. It's different from `WarningCount` in that
-// `WarningCount` return the warning count of the last executed command, so if
-// the last command is a SHOW statement, `WarningCount` return 0. On the other
-// hand, `NumWarnings` always return number of warnings(or errors if `errOnly`
+const zero = "0"
+
+// NumWarnings gets warning and error count.
 // is set).
-func (sc *StatementContext) NumWarnings(errOnly bool) uint16 {
-	var wc uint16
+func (sc *StatementContext) NumErrorWarnings() (ec, wc string) {
+	var (
+		ecNum uint16
+		wcNum int
+	)
 	sc.mu.Lock()
-	defer sc.mu.Unlock()
-	if errOnly {
-		for _, warn := range sc.mu.warnings {
-			if warn.Level == WarnLevelError {
-				wc++
-			}
-		}
+	ecNum = sc.mu.errorCount
+	wcNum = len(sc.mu.warnings)
+	sc.mu.Unlock()
+
+	if ecNum == 0 {
+		ec = zero
 	} else {
-		wc = uint16(len(sc.mu.warnings))
+		ec = strconv.Itoa(int(ecNum))
 	}
-	return wc
+
+	if wcNum == 0 {
+		wc = zero
+	} else {
+		wc = strconv.Itoa(wcNum)
+	}
+	return
 }
 
 // SetWarnings sets warnings.
 func (sc *StatementContext) SetWarnings(warns []SQLWarn) {
 	sc.mu.Lock()
 	sc.mu.warnings = warns
+	for _, w := range warns {
+		if w.Level == WarnLevelError {
+			sc.mu.errorCount++
+		}
+	}
 	sc.mu.Unlock()
 }
 
@@ -316,6 +330,7 @@ func (sc *StatementContext) AppendError(warn error) {
 	sc.mu.Lock()
 	if len(sc.mu.warnings) < math.MaxUint16 {
 		sc.mu.warnings = append(sc.mu.warnings, SQLWarn{WarnLevelError, warn})
+		sc.mu.errorCount++
 	}
 	sc.mu.Unlock()
 }
@@ -375,6 +390,7 @@ func (sc *StatementContext) ResetForRetry() {
 	sc.mu.copied = 0
 	sc.mu.touched = 0
 	sc.mu.message = ""
+	sc.mu.errorCount = 0
 	sc.mu.warnings = nil
 	sc.mu.Unlock()
 	sc.TableIDs = sc.TableIDs[:0]