--- a/store/tikv/backoff.go
+++ b/store/tikv/backoff.go
@@ -16,6 +16,7 @@ package tikv
 import (
 	"context"
 	"fmt"
+	"github.com/prometheus/client_golang/prometheus"
 	"math"
 	"math/rand"
 	"strings"
@@ -142,6 +143,37 @@ func (t backoffType) String() string {
 	return ""
 }
 
+func (t backoffType) Counter() prometheus.Counter {
+	switch t {
+	case boTiKVRPC:
+		return tikvBackoffCounterRPC
+	case BoTxnLock:
+		return tikvBackoffCounterLock
+	case boTxnLockFast:
+		return tikvBackoffCounterLockFast
+	case BoPDRPC:
+		return tikvBackoffCounterPD
+	case BoRegionMiss:
+		return tikvBackoffCounterRegionMiss
+	case BoUpdateLeader:
+		return tikvBackoffCounterUpdateLeader
+	case boServerBusy:
+		return tikvBackoffCounterServerBusy
+	}
+	return tikvBackoffCounterEmpty
+}
+
+var (
+	tikvBackoffCounterRPC          = metrics.TiKVBackoffCounter.WithLabelValues("tikvRPC")
+	tikvBackoffCounterLock         = metrics.TiKVBackoffCounter.WithLabelValues("txnLock")
+	tikvBackoffCounterLockFast     = metrics.TiKVBackoffCounter.WithLabelValues("tikvLockFast")
+	tikvBackoffCounterPD           = metrics.TiKVBackoffCounter.WithLabelValues("pdRPC")
+	tikvBackoffCounterRegionMiss   = metrics.TiKVBackoffCounter.WithLabelValues("regionMiss")
+	tikvBackoffCounterUpdateLeader = metrics.TiKVBackoffCounter.WithLabelValues("updateLeader")
+	tikvBackoffCounterServerBusy   = metrics.TiKVBackoffCounter.WithLabelValues("serverBusy")
+	tikvBackoffCounterEmpty        = metrics.TiKVBackoffCounter.WithLabelValues("")
+)
+
 func (t backoffType) TError() error {
 	switch t {
 	case boTiKVRPC:
@@ -222,7 +254,7 @@ func (b *Backoffer) Backoff(typ backoffType, err error) error {
 	default:
 	}
 
-	metrics.TiKVBackoffCounter.WithLabelValues(typ.String()).Inc()
+	typ.Counter().Inc()
 	// Lazy initialize.
 	if b.fn == nil {
 		b.fn = make(map[backoffType]func(context.Context) int)