<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

- prometheus's `WithLabelValues` will require heavy mutex/hash operation in pprof 
- opentracing stack exists in pprof even if we doesn't run sql with `trace select ...`
- grpc run with prometheus's wrapper by default but our grafana don't use them

### What is changed and how it works?

- prev with fixed label value and assgin to local at once
- remove opentracing call when we run sql without `trace sql`
- remove grpc client's prometheus's wrapper when it's not `debug mode`

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
-  sysbench point-get

run sysbench point-get with 1 tidb , 1pd, 1 unionstore(master) in my laptop

before this pr:

```
sysbench --config-file=tidb-config oltp_point_select --tables=32 --threads=32 --table-size=100000 run   
Unknown argument type: 0Unknown argument type: 0sysbench 1.1.0-c0bbed6 (using bundled LuaJIT 2.1.0-beta3)

Running the test with following options:
Number of threads: 32
Report intermediate results every 5 second(s)
Initializing random number generator from current time


Initializing worker threads...

Threads started!

[ 5s ] thds: 32 tps: 37192.07 qps: 37192.07 (r/w/o: 37192.07/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 10s ] thds: 32 tps: 36009.02 qps: 36009.02 (r/w/o: 36009.02/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 15s ] thds: 32 tps: 37087.25 qps: 37087.25 (r/w/o: 37087.25/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 32 tps: 36230.77 qps: 36230.77 (r/w/o: 36230.77/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 25s ] thds: 32 tps: 36506.36 qps: 36506.36 (r/w/o: 36506.36/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 32 tps: 36427.74 qps: 36427.74 (r/w/o: 36427.74/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 35s ] thds: 32 tps: 36623.29 qps: 36623.29 (r/w/o: 36623.29/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 40s ] thds: 32 tps: 36370.51 qps: 36370.51 (r/w/o: 36370.51/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 45s ] thds: 32 tps: 36492.56 qps: 36492.56 (r/w/o: 36492.56/0.00/0.00) lat (ms,95%): 1.76 err/s: 0.00 reconn/s: 0.00
[ 50s ] thds: 32 tps: 35857.27 qps: 35857.27 (r/w/o: 35857.27/0.00/0.00) lat (ms,95%): 1.79 err/s: 0.00 reconn/s: 0.00
[ 55s ] thds: 32 tps: 36534.97 qps: 36534.97 (r/w/o: 36534.97/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 60s ] thds: 32 tps: 36695.14 qps: 36695.14 (r/w/o: 36695.14/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 65s ] thds: 32 tps: 36360.68 qps: 36360.68 (r/w/o: 36360.68/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 70s ] thds: 32 tps: 35787.92 qps: 35787.92 (r/w/o: 35787.92/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 75s ] thds: 32 tps: 36527.31 qps: 36527.31 (r/w/o: 36527.31/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 80s ] thds: 32 tps: 36316.72 qps: 36316.72 (r/w/o: 36316.72/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 85s ] thds: 32 tps: 36236.60 qps: 36236.60 (r/w/o: 36236.60/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 90s ] thds: 32 tps: 36512.32 qps: 36512.32 (r/w/o: 36512.32/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 95s ] thds: 32 tps: 36766.68 qps: 36766.68 (r/w/o: 36766.68/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 100s ] thds: 32 tps: 37001.98 qps: 37001.98 (r/w/o: 37001.98/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 105s ] thds: 32 tps: 37152.54 qps: 37152.54 (r/w/o: 37152.54/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 110s ] thds: 32 tps: 37147.89 qps: 37147.89 (r/w/o: 37147.89/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 115s ] thds: 32 tps: 37132.60 qps: 37132.60 (r/w/o: 37132.60/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 120s ] thds: 32 tps: 36888.23 qps: 36888.23 (r/w/o: 36888.23/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 125s ] thds: 32 tps: 36449.69 qps: 36449.69 (r/w/o: 36449.69/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 130s ] thds: 32 tps: 36820.11 qps: 36820.11 (r/w/o: 36820.11/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 135s ] thds: 32 tps: 36670.81 qps: 36670.81 (r/w/o: 36670.81/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 140s ] thds: 32 tps: 36791.14 qps: 36791.14 (r/w/o: 36791.14/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 145s ] thds: 32 tps: 36319.64 qps: 36319.64 (r/w/o: 36319.64/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 150s ] thds: 32 tps: 36910.19 qps: 36910.19 (r/w/o: 36910.19/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 155s ] thds: 32 tps: 36434.97 qps: 36434.97 (r/w/o: 36434.97/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 160s ] thds: 32 tps: 36446.83 qps: 36446.83 (r/w/o: 36446.83/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 165s ] thds: 32 tps: 37179.51 qps: 37179.51 (r/w/o: 37179.51/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 170s ] thds: 32 tps: 37430.92 qps: 37430.92 (r/w/o: 37430.92/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 175s ] thds: 32 tps: 37292.76 qps: 37292.76 (r/w/o: 37292.76/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 180s ] thds: 32 tps: 37335.57 qps: 37335.57 (r/w/o: 37335.57/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 185s ] thds: 32 tps: 35354.86 qps: 35354.86 (r/w/o: 35354.86/0.00/0.00) lat (ms,95%): 1.73 err/s: 0.00 reconn/s: 0.00
[ 190s ] thds: 32 tps: 36876.84 qps: 36876.84 (r/w/o: 36876.84/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 195s ] thds: 32 tps: 37123.33 qps: 37123.33 (r/w/o: 37123.33/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 200s ] thds: 32 tps: 37001.36 qps: 37001.36 (r/w/o: 37001.36/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 205s ] thds: 32 tps: 37001.62 qps: 37001.62 (r/w/o: 37001.62/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 210s ] thds: 32 tps: 36687.51 qps: 36687.51 (r/w/o: 36687.51/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 215s ] thds: 32 tps: 36897.51 qps: 36897.51 (r/w/o: 36897.51/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 220s ] thds: 32 tps: 36460.74 qps: 36460.74 (r/w/o: 36460.74/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 225s ] thds: 32 tps: 36807.67 qps: 36807.67 (r/w/o: 36807.67/0.00/0.00) lat (ms,95%): 1.61 err/s: 0.00 reconn/s: 0.00
[ 230s ] thds: 32 tps: 36907.49 qps: 36907.49 (r/w/o: 36907.49/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 235s ] thds: 32 tps: 36627.26 qps: 36627.26 (r/w/o: 36627.26/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 240s ] thds: 32 tps: 36593.21 qps: 36593.21 (r/w/o: 36593.21/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 245s ] thds: 32 tps: 36789.74 qps: 36789.74 (r/w/o: 36789.74/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 250s ] thds: 32 tps: 36879.52 qps: 36879.52 (r/w/o: 36879.52/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 255s ] thds: 32 tps: 36554.33 qps: 36554.33 (r/w/o: 36554.33/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 260s ] thds: 32 tps: 36529.86 qps: 36529.86 (r/w/o: 36529.86/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 265s ] thds: 32 tps: 36711.03 qps: 36711.03 (r/w/o: 36711.03/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 270s ] thds: 32 tps: 36862.05 qps: 36862.05 (r/w/o: 36862.05/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 275s ] thds: 32 tps: 36332.89 qps: 36332.89 (r/w/o: 36332.89/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 280s ] thds: 32 tps: 36360.91 qps: 36360.91 (r/w/o: 36360.91/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 285s ] thds: 32 tps: 36873.78 qps: 36873.78 (r/w/o: 36873.78/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
[ 290s ] thds: 32 tps: 36196.82 qps: 36196.82 (r/w/o: 36196.82/0.00/0.00) lat (ms,95%): 1.70 err/s: 0.00 reconn/s: 0.00
[ 295s ] thds: 32 tps: 36508.55 qps: 36508.55 (r/w/o: 36508.55/0.00/0.00) lat (ms,95%): 1.67 err/s: 0.00 reconn/s: 0.00
[ 300s ] thds: 32 tps: 36547.64 qps: 36547.64 (r/w/o: 36547.64/0.00/0.00) lat (ms,95%): 1.64 err/s: 0.00 reconn/s: 0.00
SQL statistics:
    queries performed:
        read:                            10997248
        write:                           0
        other:                           0
        total:                           10997248
    transactions:                        10997248 (36656.71 per sec.)
    queries:                             10997248 (36656.71 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

Throughput:
    events/s (eps):                      36656.7142
    time elapsed:                        300.0064s
    total number of events:              10997248

Latency (ms):
         min:                                    0.14
         avg:                                    0.87
         max:                                   28.64
         95th percentile:                        1.70
         sum:                              9592300.25

Threads fairness:
    events (avg/stddev):           343664.0000/349.44
    execution time (avg/stddev):   299.7594/0.01

```

after this 

```
sysbench --config-file=tidb-config oltp_point_select --tables=32 --threads=32 --table-size=100000 run   
Unknown argument type: 0Unknown argument type: 0sysbench 1.1.0-c0bbed6 (using bundled LuaJIT 2.1.0-beta3)

Running the test with following options:
Number of threads: 32
Report intermediate results every 5 second(s)
Initializing random number generator from current time


Initializing worker threads...

Threads started!

[ 5s ] thds: 32 tps: 41333.86 qps: 41333.86 (r/w/o: 41333.86/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 10s ] thds: 32 tps: 40846.79 qps: 40846.79 (r/w/o: 40846.79/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 15s ] thds: 32 tps: 41019.58 qps: 41019.58 (r/w/o: 41019.58/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 32 tps: 40536.73 qps: 40536.73 (r/w/o: 40536.73/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 25s ] thds: 32 tps: 40711.61 qps: 40711.61 (r/w/o: 40711.61/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 32 tps: 40799.22 qps: 40799.22 (r/w/o: 40799.22/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 35s ] thds: 32 tps: 40573.79 qps: 40573.79 (r/w/o: 40573.79/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 40s ] thds: 32 tps: 40571.52 qps: 40571.52 (r/w/o: 40571.52/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 45s ] thds: 32 tps: 40942.46 qps: 40942.46 (r/w/o: 40942.46/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 50s ] thds: 32 tps: 40533.36 qps: 40533.36 (r/w/o: 40533.36/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 55s ] thds: 32 tps: 40448.16 qps: 40448.16 (r/w/o: 40448.16/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 60s ] thds: 32 tps: 40663.91 qps: 40663.91 (r/w/o: 40663.91/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 65s ] thds: 32 tps: 40557.97 qps: 40557.97 (r/w/o: 40557.97/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 70s ] thds: 32 tps: 39962.69 qps: 39962.69 (r/w/o: 39962.69/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 75s ] thds: 32 tps: 40248.64 qps: 40248.64 (r/w/o: 40248.64/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 80s ] thds: 32 tps: 40497.06 qps: 40497.06 (r/w/o: 40497.06/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 85s ] thds: 32 tps: 40080.60 qps: 40080.60 (r/w/o: 40080.60/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 90s ] thds: 32 tps: 40403.85 qps: 40403.85 (r/w/o: 40403.85/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 95s ] thds: 32 tps: 40542.22 qps: 40542.22 (r/w/o: 40542.22/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 100s ] thds: 32 tps: 40773.42 qps: 40773.42 (r/w/o: 40773.42/0.00/0.00) lat (ms,95%): 1.47 err/s: 0.00 reconn/s: 0.00
[ 105s ] thds: 32 tps: 40909.28 qps: 40909.28 (r/w/o: 40909.28/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 110s ] thds: 32 tps: 40455.26 qps: 40455.26 (r/w/o: 40455.26/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 115s ] thds: 32 tps: 40387.26 qps: 40387.26 (r/w/o: 40387.26/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 120s ] thds: 32 tps: 40977.52 qps: 40977.52 (r/w/o: 40977.52/0.00/0.00) lat (ms,95%): 1.47 err/s: 0.00 reconn/s: 0.00
[ 125s ] thds: 32 tps: 40734.28 qps: 40734.28 (r/w/o: 40734.28/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 130s ] thds: 32 tps: 40318.51 qps: 40318.51 (r/w/o: 40318.51/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 135s ] thds: 32 tps: 40345.74 qps: 40345.74 (r/w/o: 40345.74/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 140s ] thds: 32 tps: 40609.52 qps: 40609.52 (r/w/o: 40609.52/0.00/0.00) lat (ms,95%): 1.47 err/s: 0.00 reconn/s: 0.00
[ 145s ] thds: 32 tps: 40253.38 qps: 40253.38 (r/w/o: 40253.38/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 150s ] thds: 32 tps: 40720.80 qps: 40720.80 (r/w/o: 40720.80/0.00/0.00) lat (ms,95%): 1.47 err/s: 0.00 reconn/s: 0.00
[ 155s ] thds: 32 tps: 40543.30 qps: 40543.30 (r/w/o: 40543.30/0.00/0.00) lat (ms,95%): 1.47 err/s: 0.00 reconn/s: 0.00
[ 160s ] thds: 32 tps: 40050.42 qps: 40050.42 (r/w/o: 40050.42/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 165s ] thds: 32 tps: 40896.46 qps: 40896.46 (r/w/o: 40896.46/0.00/0.00) lat (ms,95%): 1.47 err/s: 0.00 reconn/s: 0.00
[ 170s ] thds: 32 tps: 39034.16 qps: 39034.16 (r/w/o: 39034.16/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 175s ] thds: 32 tps: 38965.23 qps: 38965.23 (r/w/o: 38965.23/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 180s ] thds: 32 tps: 40076.24 qps: 40076.24 (r/w/o: 40076.24/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 185s ] thds: 32 tps: 39873.56 qps: 39873.56 (r/w/o: 39873.56/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 190s ] thds: 32 tps: 40225.93 qps: 40225.93 (r/w/o: 40225.93/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 195s ] thds: 32 tps: 39978.25 qps: 39978.25 (r/w/o: 39978.25/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 200s ] thds: 32 tps: 39060.11 qps: 39060.11 (r/w/o: 39060.11/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 205s ] thds: 32 tps: 39789.40 qps: 39789.40 (r/w/o: 39789.40/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 210s ] thds: 32 tps: 40095.31 qps: 40095.31 (r/w/o: 40095.31/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 215s ] thds: 32 tps: 40090.20 qps: 40090.20 (r/w/o: 40090.20/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 220s ] thds: 32 tps: 40010.11 qps: 40010.11 (r/w/o: 40010.11/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 225s ] thds: 32 tps: 40140.50 qps: 40140.50 (r/w/o: 40140.50/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 230s ] thds: 32 tps: 39594.73 qps: 39594.73 (r/w/o: 39594.73/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 235s ] thds: 32 tps: 39474.16 qps: 39474.16 (r/w/o: 39474.16/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 240s ] thds: 32 tps: 39745.96 qps: 39745.96 (r/w/o: 39745.96/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 245s ] thds: 32 tps: 40217.78 qps: 40217.78 (r/w/o: 40217.78/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 250s ] thds: 32 tps: 39946.98 qps: 39946.98 (r/w/o: 39946.98/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 255s ] thds: 32 tps: 40228.24 qps: 40228.24 (r/w/o: 40228.24/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 260s ] thds: 32 tps: 39640.51 qps: 39640.51 (r/w/o: 39640.51/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 265s ] thds: 32 tps: 39914.15 qps: 39914.15 (r/w/o: 39914.15/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 270s ] thds: 32 tps: 39161.59 qps: 39161.59 (r/w/o: 39161.59/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 275s ] thds: 32 tps: 40124.40 qps: 40124.40 (r/w/o: 40124.40/0.00/0.00) lat (ms,95%): 1.47 err/s: 0.00 reconn/s: 0.00
[ 280s ] thds: 32 tps: 40120.86 qps: 40120.86 (r/w/o: 40120.86/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 285s ] thds: 32 tps: 39865.57 qps: 39865.57 (r/w/o: 39865.57/0.00/0.00) lat (ms,95%): 1.52 err/s: 0.00 reconn/s: 0.00
[ 290s ] thds: 32 tps: 39086.00 qps: 39086.00 (r/w/o: 39086.00/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
[ 295s ] thds: 32 tps: 40042.72 qps: 40042.72 (r/w/o: 40042.72/0.00/0.00) lat (ms,95%): 1.50 err/s: 0.00 reconn/s: 0.00
[ 300s ] thds: 32 tps: 38923.58 qps: 38923.58 (r/w/o: 38923.58/0.00/0.00) lat (ms,95%): 1.55 err/s: 0.00 reconn/s: 0.00
SQL statistics:
    queries performed:
        read:                            12063493
        write:                           0
        other:                           0
        total:                           12063493
    transactions:                        12063493 (40210.99 per sec.)
    queries:                             12063493 (40210.99 per sec.)
    ignored errors:                      0      (0.00 per sec.)
    reconnects:                          0      (0.00 per sec.)

Throughput:
    events/s (eps):                      40210.9893
    time elapsed:                        300.0049s
    total number of events:              12063493

Latency (ms):
         min:                                    0.12
         avg:                                    0.80
         max:                                   27.31
         95th percentile:                        1.52
         sum:                              9591690.90

Threads fairness:
    events (avg/stddev):           376984.1562/300.90
    execution time (avg/stddev):   299.7403/0.01

```

Code changes

 - implement change

Side effects

 - Increased code complexity

Related changes

 - N/A

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/pingcap/tidb/10158)
<!-- Reviewable:end -->